+++ generators/app/generate.php
@@ -27,7 +27,8 @@ class AppGenerator extends \Dxw\Whippet\WhippetGenerator {
     }
 
     // Make the target dir a git repo, if it isn't already
-    if(!(new \Dxw\Whippet\Git\Git($this->target_dir))->is_repo()) {
+    $target_repo = new \Dxw\Whippet\Git\Git($this->target_dir);
+    if(!$target_repo->is_repo()) {
       \Dxw\Whippet\Git\Git::init($this->target_dir);
     }
 
@@ -47,6 +48,12 @@ class AppGenerator extends \Dxw\Whippet\WhippetGenerator {
     exec("chmod 0755 " . $this->target_dir . "/setup/*");
     exec("chmod 0755 " . $this->target_dir . "/script/*");
     exec("chmod 0755 " . $this->target_dir . "/bin/*");
+
+    // Whippet deploy requires at least one commit in the repo.
+    if (!$target_repo->current_commit()) {
+      $target_repo->add("--all");
+      $target_repo->commit("Initial commit from Whippet");
+    }
 
     echo "New whippet app successfully generated at {$this->target_dir} \n";
   }