+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+- `whippet generate app` generates an app which can be deployed by Whippet.
+
 ### Added
 - Support for PHP 8
 - `whippet generate plugin` to generate a plugin based on our template repo https://github.com/dxw/wordpress-plugin-template/